完善代码1

+++ client-pc/src/main/java/com/sys/clientpc/netty/listeners/CloseScreenListener.java
@@ -1,5 +1,9 @@
 package com.sys.clientpc.netty.listeners;
 
+import ch.qos.logback.core.db.dialect.MySQLDialect;
+import com.sys.clientpc.netty.SocketIOClient;
+import com.sys.clientpc.utils.SpringContextHolder;
+import com.sys.clientpc.webcam.MyWebcam;
 import io.socket.emitter.Emitter;
 import lombok.extern.slf4j.Slf4j;
 
@@ -14,34 +18,42 @@ import java.io.IOException;
  **/
 @Slf4j
 public class CloseScreenListener implements Emitter.Listener   {
+    SocketIOClient socketIOClient  = SocketIOClient.getInstance();
     @Override
     public void call(Object... o) {
-        log.debug("接收到服务端指令pc-close");
-        String str = String.valueOf(o);
+        if(o==null){
+            return;
+        }
+        String str = (String) o[0];
+        log.info("接收到服务端指令 pc-close "+str);
         try {
+            String res = "1";
             switch (str) {
-                case "local":
+                case "lock_screen":
                     Runtime.getRuntime().exec("RunDll32.exe user32.dll,LockWorkStation");
-//                case "local":
-//                    Runtime.getRuntime().exec("shutdown -s -f -t 00");
-//                case "local":
-//                    Runtime.getRuntime().exec("RunDll32.exe user32.dll,LockWorkStation");
-//                case "local":
-//                    Runtime.getRuntime().exec("RunDll32.exe user32.dll,LockWorkStation");
-//                case "local":
-//                    Runtime.getRuntime().exec("RunDll32.exe user32.dll,LockWorkStation");
+                    break;
+                case "shutdown_now":
+                    Runtime.getRuntime().exec("shutdown -s -f -t 00");
+                    break;
+                case "shutdown60":
+                    Runtime.getRuntime().exec("shutdown -s -f -t 60");
+                    break;
+                case "cancel_shutdown":
+                    Runtime.getRuntime().exec("shutdown -a");
+                    break;
+                case "sleep":
+                    Runtime.getRuntime().exec("shutdown -h");
+                    break;
+                case "take_photo":
+                    res = MyWebcam.getImageBase64();
+                    break;
             }
             //向服务端响应结果
+            socketIOClient.getSocket().emit("pc-close-res",res);
         } catch (IOException e) {
             e.printStackTrace();
+            socketIOClient.getSocket().emit("pc-close-res","0");
         }
     }
 
-    public static void main(String[] args) {
-        try {
-            Runtime.getRuntime().exec("shutdown -s -f -t 00");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
 }
